fix(quant): min_seconds_between_orders 支持 float（亚秒级 cooldown）
原 type=int 导致 0.5 这种亚秒值无法配置——0 是没冷却（散户挤兑下 backend
可能限流 / 单笔 cap 仍兜底），1 又错失短时多笔成交的机会。

改成 float 后可以填 0.3 / 0.5 等折中值。trader.py 的 int() 转换同步改成
float()。risk.py 的 time.monotonic() - last < cfg 浮点比较本来就工作。

13 risk+broker 测试无回归（原测试 0 / 3 都是合法 float）。

+++ quant/thccb_quant/broker/risk.py
@@ -13,7 +13,7 @@ class RiskConfig:
     daily_loss_cap_cny: Decimal
     daily_turnover_cap_cny: Decimal
     max_slippage_bps: int
-    min_seconds_between_orders: int
+    min_seconds_between_orders: float  # 支持亚秒粒度（如 0.5）应对散户挤兑
 
 
 class RiskGuard:
